Day 02
*Sign Up Process

It assigned IDs to input fields. After that, the script file created variables for data that transfer from input fields in the signup page design.

To transfer data from the script file to the signup process, I used Ajax Technology. In the script file a new form of data to transfer data through the POST method.

After that in the Signup process file, We validate data that comes from input fields.

->empty($variable)
this is used to check that a variable is empty or not

->strlen($variable)
this is used to get the character length of the variable

->preg_match("data Structure" , $vaiable)
this is used to check whether data is in the structure that we need it to be

After validating data using the function we created in the connection.php file, executed the insert query to add user data to the database, and sent the response as a success.

In a script file, it checks whether the response is Success or not

If it was a success we send a success alert and redirect the user to the sign in form.

If it isn't, sends an alert to the user that contains the problem