Day 03
* Sign in form input validation
Same as before, used the empty function and str_len to validate the input fields.

*Sign-in process
Same as in the sign-up form, firstly assigned IDs for each input field. After that create a function called SignUp() and set it to click on the sign up button.
In this function, we assign data that comes from input fields to variables and send those to a PHP file that contains the sign-in process by the POST method.
in-process file executes a search function and checks whether the data is valid.
suppose it wasn't alerted the user that the data was not valid. If it was assigned data that fetches from the database to the session and redirects the user to the home page.

*Remember me option
If this option is ticked by the user in the process file we create two cookies that are valid for one year time. One is for an email address that the user entered while signing in and the other one is for a password. On the design side first, we check whether there is a cookie that carries that user email or password. If there were any we assign them as the value of input files so the user can easily log in to the account that he owns

*Forgot password modal design
If the user has forgotten his password there is an option to reset the password
For this process, we used a bootstrap model and called onclick at anchor tag.
for the input fields, use input groups and the created button to view or not the password characters. To that, we used input type, and using javascript it changed from password to text when it clicked.* Sign in form input validation
Same as before, used the empty function and str_len to validate the input fields.

*Sign-in process
Same as in the sign-up form, firstly assigned IDs for each input field. After that create a function called SignUp() and set it to click on the sign up button.
In this function, we assign data that comes from input fields to variables and send those to a PHP file that contains the sign-in process by the POST method.
in-process file executes a search function and checks whether the data is valid.
suppose it wasn't alerted the user that the data was not valid. If it was assigned data that fetches from the database to the session and redirects the user to the home page.

*Remember me option
If this option is ticked by the user in the process file we create two cookies that are valid for one year time. One is for an email address that the user entered while signing in and the other one is for a password. On the design side first, we check whether there is a cookie that carries that user email or password. If there were any we assign them as the value of input files so the user can easily log in to the account that he owns

*Forgot password modal design
If the user has forgotten his password there is an option to reset the password
For this process, we used a bootstrap model and called onclick at anchor tag.
for the input fields, use input groups and the created button to view or not the password characters. To that, we used input type, and using javascript it changed from password to text when it clicked.