Day 05
*Header User Name Section

In header section we check whether user is logged or not by web session. If use has been logged it displays the first name of user with greeting .If not it displays user to sign in.

*Category selector
In the search option we can a selector to select product category. In it, firstly we executed a search query and get the numbers of rows that in category table. After that it runs for loop to display selector option that contains in  data base.

*Home category sections
In home page there section to every category that shows latest four products in it. In this section also runs a for loop to generate category section and each one has own for loop that executes a search query and filter lastly added four products and display them as product cards.

also there is simple design when there ins no product to view😋

*Sign out Process
We have a sing out section next to the user greeting. In that section it calls a sign out function onclick the sign out text.  In that process firstly checks whether is there any session runs in web page or not. If it is, this process will destroy the session. Because of that user would sign out from the web app.